.history/requirements_20250817132947.txt
fastapi
uvicorn
torch
torchvision
Pillow
mlflow
scikit-learn
numpy
opencv-python-headless
gradio
